code_scout/pyproject.toml
[tool.poetry]
name = "code_scout"
version = "0.1.0"
description = "AI Code Reviewer"
authors = ["Andrzej Chmielewski <andrzejchm@gmail.com>"]
license = "Business Source License 1.1"
packages = [
    { include = "core", from = "src" },
    { include = "cli", from = "src" },
]

[tool.poetry.scripts]
codescout = "cli.main:app"

[tool.poetry.dependencies]
python = ">=3.13, <4.0"
rich = ">=14.0.0,<15.0.0"
langchain = ">=0.3.25,<0.4.0"
pydantic = ">=2.11.5,<3.0.0"
gitpython = ">=3.1.44,<4.0.0"
unidiff = ">=0.7.5,<0.8.0"
typer = "^0.16.0"
python-dotenv = "^1.1.1"
langchain-community = "^0.3.27"
langchain-openai = "^0.3.27"
langchain-anthropic = "^0.3.17"
pygithub = "^2.2.0"
questionary = "^2.1.0"
langgraph = "^0.6.5"
tree-sitter = "^0.25.1"
gitignore-parser = "^0.1.12"
tree-sitter-language-pack = "^0.9.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.10"
pytest = ">=8.3.5,<9.0.0"
pytest-cov = ">=6.1.1,<7.0.0"
basedpyright = "^1.31.3"
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.basedpyright]
include = ["src", "tests"]
venvPath = "."
venv = ".venv"
exclude = ["**/__pycache__"]
typeCheckingMode = "recommended"
reportMissingImports = "error"
reportMissingTypeStubs = false
reportExplicitAny = false
reportUnknownArgumentType= false
reportUnknownMemberType= false
reportCallInDefaultInitializer=false
reportAny = false
pythonVersion = "3.13"
pythonPlatform = "All"

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "SIM", "ARG", "PL", "RUF", "ANN"]
ignore = ["ANN204", "ANN401"]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = ["PLR2004"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["scraper"]
